fix(stringToNumber): Se corrigio la funcion de formatear numeros

+++ src/js/utils/formatUtils.js
@@ -10,7 +10,7 @@ export const formatNumber = (number) => { //Cambiar nombre a "formatNumberToStri
     return number.toLocaleString('en', { minimumFractionDigits: 2, maximumFractionDigits: 2 });
 }
 export const stringToNumber = (string) => { //Cambiar nombre a "formatStringToNumber"
-    return parseInt(string.replace(/[^0-9.-]/g, ""));
+    return parseFloat(string.replace(/[^0-9.-]/g, ""));
 }
 export const detectSubdomain = () => {
     let hostname = window.location.hostname;
